Fix Mirror Maker 2 connector reconciliation race condition
